backend/requirements.txt
flask
faiss-cpu
numpy
sentence-transformers
flask-cors
flask-restful